+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.1.3] - 2026-09-23
+
+### Added
+- Added separate AWS SQS and SNS-to-SQS examples.
+- Standardized public provider names to `AWS` and `RABBITMQ`.
+
 ## [0.1.2] - 2026-09-23
 
 ### Fixed